fix: remove internationalization for now as its causing too many issues
