+++ src/app/(auth)/(center)/layout.tsx
@@ -1,0 +1,11 @@
+export default function CenteredLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <div className='flex min-h-screen items-center justify-center'>
+      {children}
+    </div>
+  )
+}
