--- src/app/[locale]/(auth)/(center)/layout.tsx
@@ -1,15 +1,0 @@
-import { setRequestLocale } from 'next-intl/server'
-
-export default async function CenteredLayout(props: {
-  children: React.ReactNode
-  params: Promise<{ locale: string }>
-}) {
-  const { locale } = await props.params
-  setRequestLocale(locale)
-
-  return (
-    <div className='flex min-h-screen items-center justify-center'>
-      {props.children}
-    </div>
-  )
-}
