--- src/app/[locale]/(auth)/(center)/sign-in/[[...sign-in]]/page.tsx
@@ -1,32 +1,0 @@
-import type { Metadata } from 'next'
-
-import { SignIn } from '@clerk/nextjs'
-import { getTranslations, setRequestLocale } from 'next-intl/server'
-
-import { getI18nPath } from '@/utils/Helpers'
-
-type ISignInPageProps = {
-  params: Promise<{ locale: string }>
-}
-
-export async function generateMetadata(
-  props: ISignInPageProps
-): Promise<Metadata> {
-  const { locale } = await props.params
-  const t = await getTranslations({
-    locale,
-    namespace: 'SignIn',
-  })
-
-  return {
-    title: t('meta_title'),
-    description: t('meta_description'),
-  }
-}
-
-export default async function SignInPage(props: ISignInPageProps) {
-  const { locale } = await props.params
-  setRequestLocale(locale)
-
-  return <SignIn path={getI18nPath('/sign-in', locale)} />
-}
