--- src/app/[locale]/(auth)/(center)/sign-up/[[...sign-up]]/page.tsx
@@ -1,32 +1,0 @@
-import type { Metadata } from 'next'
-
-import { SignUp } from '@clerk/nextjs'
-import { getTranslations, setRequestLocale } from 'next-intl/server'
-
-import { getI18nPath } from '@/utils/Helpers'
-
-type ISignUpPageProps = {
-  params: Promise<{ locale: string }>
-}
-
-export async function generateMetadata(
-  props: ISignUpPageProps
-): Promise<Metadata> {
-  const { locale } = await props.params
-  const t = await getTranslations({
-    locale,
-    namespace: 'SignUp',
-  })
-
-  return {
-    title: t('meta_title'),
-    description: t('meta_description'),
-  }
-}
-
-export default async function SignUpPage(props: ISignUpPageProps) {
-  const { locale } = await props.params
-  setRequestLocale(locale)
-
-  return <SignUp path={getI18nPath('/sign-up', locale)} />
-}
