--- src/app/[locale]/(auth)/api/users/route.ts
@@ -1,128 +1,0 @@
-import type { NextRequest} from 'next/server';
-
-import { auth } from '@clerk/nextjs/server'
-import { eq } from 'drizzle-orm'
-import { NextResponse } from 'next/server'
-
-import { db } from '@/libs/DB'
-import { users } from '@/models/Schema'
-
-// GET /api/users - Get current user profile
-export async function GET() {
-  try {
-    const { userId } = await auth()
-
-    if (!userId) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
-    }
-
-    const user = await db.query.users.findFirst({
-      where: eq(users.clerkId, userId),
-    })
-
-    if (!user) {
-      return NextResponse.json({ error: 'User not found' }, { status: 404 })
-    }
-
-    return NextResponse.json(user)
-  } catch (error) {
-    console.error('Error fetching user:', error)
-    return NextResponse.json(
-      { error: 'Internal server error' },
-      { status: 500 }
-    )
-  }
-}
-
-// POST /api/users - Create or update user profile
-export async function POST(request: NextRequest) {
-  try {
-    const { userId } = await auth()
-
-    if (!userId) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
-    }
-
-    const body = await request.json()
-    const { email, firstName, lastName, avatarUrl } = body
-
-    // Check if user already exists
-    const existingUser = await db.query.users.findFirst({
-      where: eq(users.clerkId, userId),
-    })
-
-    if (existingUser) {
-      // Update existing user
-      const updatedUser = await db
-        .update(users)
-        .set({
-          email,
-          firstName,
-          lastName,
-          avatarUrl,
-          updatedAt: new Date(),
-        })
-        .where(eq(users.clerkId, userId))
-        .returning()
-
-      return NextResponse.json(updatedUser[0])
-    } else {
-      // Create new user
-      const newUser = await db
-        .insert(users)
-        .values({
-          clerkId: userId,
-          email,
-          firstName,
-          lastName,
-          avatarUrl,
-        })
-        .returning()
-
-      return NextResponse.json(newUser[0], { status: 201 })
-    }
-  } catch (error) {
-    console.error('Error creating/updating user:', error)
-    return NextResponse.json(
-      { error: 'Internal server error' },
-      { status: 500 }
-    )
-  }
-}
-
-// PUT /api/users - Update user profile
-export async function PUT(request: NextRequest) {
-  try {
-    const { userId } = await auth()
-
-    if (!userId) {
-      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
-    }
-
-    const body = await request.json()
-    const { firstName, lastName, avatarUrl } = body
-
-    const updatedUser = await db
-      .update(users)
-      .set({
-        firstName,
-        lastName,
-        avatarUrl,
-        updatedAt: new Date(),
-      })
-      .where(eq(users.clerkId, userId))
-      .returning()
-
-    if (updatedUser.length === 0) {
-      return NextResponse.json({ error: 'User not found' }, { status: 404 })
-    }
-
-    return NextResponse.json(updatedUser[0])
-  } catch (error) {
-    console.error('Error updating user:', error)
-    return NextResponse.json(
-      { error: 'Internal server error' },
-      { status: 500 }
-    )
-  }
-}
