--- src/app/[locale]/(auth)/layout.tsx
@@ -1,44 +1,0 @@
-import { ClerkProvider } from '@clerk/nextjs'
-import { setRequestLocale } from 'next-intl/server'
-
-import { routing } from '@/libs/I18nRouting'
-import { ClerkLocalizations } from '@/utils/AppConfig'
-
-export default async function AuthLayout(props: {
-  children: React.ReactNode
-  params: Promise<{ locale: string }>
-}) {
-  const { locale } = await props.params
-  setRequestLocale(locale)
-
-  const clerkLocale =
-    ClerkLocalizations.supportedLocales[locale] ??
-    ClerkLocalizations.defaultLocale
-  let signInUrl = '/sign-in'
-  let signUpUrl = '/sign-up'
-  let dashboardUrl = '/dashboard'
-  let afterSignOutUrl = '/'
-
-  if (locale !== routing.defaultLocale) {
-    signInUrl = `/${locale}${signInUrl}`
-    signUpUrl = `/${locale}${signUpUrl}`
-    dashboardUrl = `/${locale}${dashboardUrl}`
-    afterSignOutUrl = `/${locale}${afterSignOutUrl}`
-  }
-
-  return (
-    <ClerkProvider
-      localization={clerkLocale}
-      signInUrl={signInUrl}
-      signUpUrl={signUpUrl}
-      signInFallbackRedirectUrl={dashboardUrl}
-      signUpFallbackRedirectUrl={dashboardUrl}
-      afterSignOutUrl={afterSignOutUrl}
-      appearance={{
-        cssLayerName: 'clerk', // Ensure Clerk is compatible with Tailwind CSS v4
-      }}
-    >
-      {props.children}
-    </ClerkProvider>
-  )
-}
