+++ src/app/api/users/route.ts
@@ -11,15 +11,32 @@ import { users } from '@/models/Schema'
 // GET /api/users - Get current user profile
 export async function GET() {
   try {
+    console.warn('=== SIMPLE LOG TEST ===')
+    logger.info('GET /api/users called')
+
     const { userId } = await auth()
+    logger.info('Auth result', {
+      userId,
+      userIdType: typeof userId,
+      userIdLength: userId?.length,
+    })
 
     if (!userId) {
       logger.warn('Unauthorized access attempt to /api/users')
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
     }
 
+    logger.info('Attempting database query', {
+      userId,
+      queryClause: `users.clerkId = '${userId}'`,
+    })
     const user = await db.query.users.findFirst({
       where: eq(users.clerkId, userId),
+    })
+    logger.info('Database query result', {
+      userFound: !!user,
+      foundUserId: user?.clerkId,
+      foundEmail: user?.email,
     })
 
     if (!user) {
@@ -30,7 +47,11 @@ export async function GET() {
     logger.debug('User profile retrieved', { clerkId: userId })
     return NextResponse.json(user)
   } catch (error) {
-    logger.error('Error fetching user profile', { error, clerkId: 'unknown' })
+    logger.error('Error fetching user profile', {
+      error: error instanceof Error ? error.message : String(error),
+      stack: error instanceof Error ? error.stack : undefined,
+      clerkId: 'unknown',
+    })
     return NextResponse.json(
       { error: 'Internal server error' },
       { status: 500 }
