+++ src/components/atoms/controls/index.ts
@@ -1,3 +1,2 @@
 export { BackButton } from './BackButton'
 export { Button } from './Button'
-export { LocaleSwitcher } from './LocaleSwitcher'
