--- src/libs/I18nNavigation.ts
@@ -1,5 +1,0 @@
-import { createNavigation } from 'next-intl/navigation'
-
-import { routing } from './I18nRouting'
-
-export const { usePathname } = createNavigation(routing)
