--- src/libs/I18nRouting.ts
@@ -1,9 +1,0 @@
-import { defineRouting } from 'next-intl/routing'
-
-import { AppConfig } from '@/utils/AppConfig'
-
-export const routing = defineRouting({
-  locales: AppConfig.locales,
-  localePrefix: AppConfig.localePrefix,
-  defaultLocale: AppConfig.defaultLocale,
-})
