+++ src/middleware.ts
@@ -2,28 +2,11 @@ import type { NextFetchEvent, NextRequest } from 'next/server'
 
 import { detectBot } from '@arcjet/next'
 import { clerkMiddleware, createRouteMatcher } from '@clerk/nextjs/server'
-import createMiddleware from 'next-intl/middleware'
 import { NextResponse } from 'next/server'
 
 import arcjet from '@/libs/Arcjet'
 
-import { routing } from './libs/I18nRouting'
-
-const handleI18nRouting = createMiddleware(routing)
-
-const isProtectedRoute = createRouteMatcher([
-  '/dashboard(.*)',
-  '/:locale/dashboard(.*)',
-])
-
-const isAuthPage = createRouteMatcher([
-  '/sign-in(.*)',
-  '/:locale/sign-in(.*)',
-  '/sign-up(.*)',
-  '/:locale/sign-up(.*)',
-])
-
-const isApiRoute = (pathname: string) => pathname.startsWith('/api/')
+const isProtectedRoute = createRouteMatcher(['/dashboard(.*)'])
 
 // Improve security with Arcjet
 const aj = arcjet.withRule(
@@ -43,12 +26,7 @@ export default async function middleware(
   request: NextRequest,
   event: NextFetchEvent
 ) {
-  const { pathname } = request.nextUrl
-
-  // Skip middleware for API routes - let Clerk handle auth directly
-  if (isApiRoute(pathname)) {
-    return NextResponse.next()
-  }
+  // Note: API routes need Clerk middleware for auth() to work
 
   // Verify the request with Arcjet
   // Use `process.env` instead of Env to reduce bundle size in middleware
@@ -60,25 +38,19 @@ export default async function middleware(
     }
   }
 
-  // Clerk keyless mode doesn't work with i18n, this is why we need to run the middleware conditionally
-  if (isAuthPage(request) || isProtectedRoute(request)) {
-    return clerkMiddleware(async (auth, req) => {
-      if (isProtectedRoute(req)) {
-        const locale =
-          req.nextUrl.pathname.match(/(\/.*)\/dashboard/)?.at(1) ?? ''
+  // Run Clerk middleware for ALL routes to enable auth() in API routes
+  return clerkMiddleware(async (auth, req) => {
+    // Only protect dashboard routes
+    if (isProtectedRoute(req)) {
+      const signInUrl = new URL('/sign-in', req.url)
 
-        const signInUrl = new URL(`${locale}/sign-in`, req.url)
+      await auth.protect({
+        unauthenticatedUrl: signInUrl.toString(),
+      })
+    }
 
-        await auth.protect({
-          unauthenticatedUrl: signInUrl.toString(),
-        })
-      }
-
-      return handleI18nRouting(request)
-    })(request, event)
-  }
-
-  return handleI18nRouting(request)
+    return NextResponse.next()
+  })(request, event)
 }
 
 export const config = {
