+++ src/templates/BaseTemplate.test.tsx
@@ -1,9 +1,6 @@
 import { page } from '@vitest/browser/context'
-import { NextIntlClientProvider } from 'next-intl'
 import { describe, expect, it } from 'vitest'
 import { render } from 'vitest-browser-react'
-
-import messages from '@/locales/en.json'
 
 import { BaseTemplate } from './BaseTemplate'
 
@@ -11,19 +8,17 @@ describe('Base template', () => {
   describe('Render method', () => {
     it('should have 3 menu items', () => {
       render(
-        <NextIntlClientProvider locale='en' messages={messages}>
-          <BaseTemplate
-            leftNav={
-              <>
-                <li>link 1</li>
-                <li>link 2</li>
-                <li>link 3</li>
-              </>
-            }
-          >
-            {null}
-          </BaseTemplate>
-        </NextIntlClientProvider>
+        <BaseTemplate
+          leftNav={
+            <>
+              <li>link 1</li>
+              <li>link 2</li>
+              <li>link 3</li>
+            </>
+          }
+        >
+          {null}
+        </BaseTemplate>
       )
 
       const menuItemList = page.getByRole('listitem')
@@ -32,11 +27,7 @@ describe('Base template', () => {
     })
 
     it('should have a link to support creativedesignsguru.com', () => {
-      render(
-        <NextIntlClientProvider locale='en' messages={messages}>
-          <BaseTemplate leftNav={<li>1</li>}>{null}</BaseTemplate>
-        </NextIntlClientProvider>
-      )
+      render(<BaseTemplate leftNav={<li>1</li>}>{null}</BaseTemplate>)
 
       const copyrightSection = page.getByText(/© Copyright/)
       const copyrightLink = copyrightSection.getByRole('link')
