+++ src/utils/AppConfig.ts
@@ -1,24 +1,10 @@
-import type { LocalizationResource } from '@clerk/types'
-import type { LocalePrefixMode } from 'next-intl/routing'
-
-import { enUS, frFR } from '@clerk/localizations'
-
-const localePrefix: LocalePrefixMode = 'as-needed'
+import { enUS } from '@clerk/localizations'
 
 // FIXME: Update this configuration file based on your project information
 export const AppConfig = {
-  name: 'Nextjs Starter',
-  locales: ['en', 'fr'],
-  defaultLocale: 'en',
-  localePrefix,
-}
-
-const supportedLocales: Record<string, LocalizationResource> = {
-  en: enUS,
-  fr: frFR,
+  name: 'c0d3ster',
 }
 
 export const ClerkLocalizations = {
   defaultLocale: enUS,
-  supportedLocales,
 }
