+++ src/utils/Helpers.ts
@@ -1,5 +1,3 @@
-import { routing } from '@/libs/I18nRouting'
-
 export const getBaseUrl = () => {
   if (process.env.NEXT_PUBLIC_APP_URL) {
     return process.env.NEXT_PUBLIC_APP_URL
@@ -17,14 +15,6 @@ export const getBaseUrl = () => {
   }
 
   return 'http://localhost:3000'
-}
-
-export const getI18nPath = (url: string, locale: string) => {
-  if (locale === routing.defaultLocale) {
-    return url
-  }
-
-  return `/${locale}${url}`
 }
 
 export const isServer = () => {
